вынесла интерфейся в отдельные папки
