+++ src/entities/cardItem/cardProduct.css
