+++ src/entities/cardItem/interfaces/interfaces.ts
@@ -1,0 +1,7 @@
+export interface CardItemProps {
+    title: string;
+    price: number;
+    images: Array<string>;
+    id: number;
+    category: number;
+  }
