--- src/entities/cardProduct/index.ts
@@ -1,1 +1,0 @@
-export {CardProduct} from './CardProduct'
