+++ src/features/catalog/api/index.ts
@@ -1,0 +1,1 @@
+export * from './catalogApi'
