+++ src/features/hits/interfaces/interfaces.ts
@@ -1,0 +1,7 @@
+export interface HitState {
+    title: string;
+    price: number;
+    images: Array<string>;
+    id: number;
+    category: number ;
+  }
